Revert "ADD: Manage Wallets Sort By"
This reverts commit b0dbe0966d76f70d9b2097f8af896b8864ca5728.
