+++ loc/en.json
@@ -383,7 +383,6 @@
     "add_bitcoin_explain": "Simple and powerful Bitcoin wallet",
     "add_create": "Create",
     "total_balance": "Total Balance",
-    "balance": "Balance",
     "add_entropy_reset_title": "Reset Entropy",
     "add_entropy_reset_message": "Changing the wallet type will reset the current entropy. Do you want to proceed?",
     "add_entropy": "Entropy",
@@ -474,8 +473,6 @@
     "no_ln_wallet_error": "Before paying a Lightning invoice, you must first add a Lightning wallet.",
     "looks_like_bip38": "This looks like a password-protected private key (BIP38).",
     "manage_title": "Manage Wallets",
-    "sort_by_order": "Order",
-    "sort_by_property": "Property",
     "no_results_found": "No results found.",
     "please_continue_scanning": "Please continue scanning.",
     "select_no_bitcoin": "There are currently no Bitcoin wallets available.",
