+++ screen/wallets/ManageWallets.tsx
@@ -1,5 +1,5 @@
 import React, { useEffect, useLayoutEffect, useReducer, useCallback, useMemo, useRef } from 'react';
-import { StyleSheet, View, TouchableOpacity, Image, Text, Alert, I18nManager, Animated, LayoutAnimation } from 'react-native';
+import { StyleSheet, TouchableOpacity, Image, Text, Alert, I18nManager, Animated, LayoutAnimation } from 'react-native';
 import {
   NestableScrollContainer,
   ScaleDecorator,
@@ -25,21 +25,10 @@ import prompt from '../../helpers/prompt';
 import HeaderRightButton from '../../components/HeaderRightButton';
 import { ManageWalletsListItem } from '../../components/ManageWalletsListItem';
 import { useSettings } from '../../hooks/context/useSettings';
-import { CommonToolTipActions } from '../../typings/CommonToolTipActions';
-import MoreOptionsButton from '../../components/icons/MoreOptionsButton';
-import { Action } from '../../components/types';
 
 enum ItemType {
   WalletSection = 'wallet',
   TransactionSection = 'transaction',
-}
-
-enum SortOption {
-  Balance = 'balance',
-  Label = 'label',
-  MostRecent = 'mostRecent',
-  ASC = 'asc',
-  DESC = 'desc',
 }
 
 interface WalletItem {
@@ -61,7 +50,6 @@ const SET_FILTERED_ORDER = 'SET_FILTERED_ORDER';
 const SET_TEMP_ORDER = 'SET_TEMP_ORDER';
 const REMOVE_WALLET = 'REMOVE_WALLET';
 const SAVE_CHANGES = 'SAVE_CHANGES';
-const SET_CURRENT_SORT = 'SET_CURRENT_SORT';
 
 interface SaveChangesAction {
   type: typeof SAVE_CHANGES;
@@ -98,20 +86,14 @@ interface RemoveWalletAction {
   payload: string; // Wallet ID
 }
 
-interface SetCurrentSortAction {
-  type: typeof SET_CURRENT_SORT;
-  payload: SortOption;
-}
-
-type ReducerAction =
+type Action =
   | SetSearchQueryAction
   | SetIsSearchFocusedAction
   | SetInitialOrderAction
   | SetFilteredOrderAction
   | SetTempOrderAction
   | SaveChangesAction
-  | RemoveWalletAction
-  | SetCurrentSortAction;
+  | RemoveWalletAction;
 
 interface State {
   searchQuery: string;
@@ -120,7 +102,6 @@ interface State {
   tempOrder: Item[];
   wallets: TWallet[];
   txMetadata: TTXMetadata;
-  currentSort: SortOption;
 }
 
 const initialState: State = {
@@ -130,14 +111,13 @@ const initialState: State = {
   tempOrder: [],
   wallets: [],
   txMetadata: {},
-  currentSort: SortOption.Balance,
 };
 
 const deepCopyWallets = (wallets: TWallet[]): TWallet[] => {
   return wallets.map(wallet => Object.assign(Object.create(Object.getPrototypeOf(wallet)), wallet));
 };
 
-const reducer = (state: State, action: ReducerAction): State => {
+const reducer = (state: State, action: Action): State => {
   switch (action.type) {
     case SET_SEARCH_QUERY:
       return { ...state, searchQuery: action.payload };
@@ -201,10 +181,8 @@ const reducer = (state: State, action: ReducerAction): State => {
         tempOrder: updatedOrder,
       };
     }
-    case SET_CURRENT_SORT:
-      return { ...state, currentSort: action.payload };
     default:
-      throw new Error(`Unhandled action type: ${(action as ReducerAction).type}`);
+      throw new Error(`Unhandled action type: ${(action as Action).type}`);
   }
 };
 
@@ -285,124 +263,9 @@ const ManageWallets: React.FC = () => {
     [goBack, closeImage],
   );
 
-  const moreOptionsActions = useMemo((): Action[] | Action[][] => {
-    return [
-      {
-        id: 'sort_by_menu',
-        text: loc.cc.sort_by,
-        subactions: [
-          {
-            id: 'sort_by_order',
-            displayInline: true,
-            text: loc.wallets.sort_by_order,
-            keepsMenuPresented: true,
-            subactions: [
-              { ...CommonToolTipActions.SortASC, menuState: state.currentSort === SortOption.ASC },
-              { ...CommonToolTipActions.SortDESC, menuState: state.currentSort === SortOption.DESC },
-            ],
-          },
-          {
-            id: 'sort_by_wallet',
-            displayInline: true,
-            text: loc.wallets.sort_by_property,
-            subactions: [
-              {
-                ...CommonToolTipActions.SortBalance,
-                menuState: state.currentSort === SortOption.Balance,
-                disabled: state.currentSort === SortOption.Balance,
-              },
-              {
-                ...CommonToolTipActions.SortLabel,
-                menuState: state.currentSort === SortOption.Label,
-                disabled: state.currentSort === SortOption.Label,
-              },
-              {
-                ...CommonToolTipActions.MostRecentTransaction,
-                menuState: state.currentSort === SortOption.MostRecent,
-                disabled: state.currentSort === SortOption.MostRecent,
-              },
-            ],
-          },
-          { ...CommonToolTipActions.Reset },
-        ],
-      },
-    ];
-  }, [state.currentSort]);
-
-  const moreOptionsOnPressMenuItem = useCallback(
-    (id: string) => {
-      LayoutAnimation.configureNext(LayoutAnimation.Presets.easeInEaseOut);
-      switch (id) {
-        case CommonToolTipActions.SortASC.id: {
-          dispatch({ type: SET_CURRENT_SORT, payload: SortOption.ASC });
-          const sortedWallets = state.tempOrder
-            .filter((item): item is WalletItem => item.type === ItemType.WalletSection)
-            .sort((a, b) => a.data.getLabel()!.localeCompare(b.data.getLabel()!));
-          dispatch({ type: SET_TEMP_ORDER, payload: sortedWallets });
-          break;
-        }
-        case CommonToolTipActions.SortDESC.id: {
-          dispatch({ type: SET_CURRENT_SORT, payload: SortOption.DESC });
-          const sortedWallets = state.tempOrder
-            .filter((item): item is WalletItem => item.type === ItemType.WalletSection)
-            .sort((a, b) => b.data.getLabel()!.localeCompare(a.data.getLabel()!));
-          dispatch({ type: SET_TEMP_ORDER, payload: sortedWallets });
-          break;
-        }
-        case CommonToolTipActions.SortBalance.id: {
-          dispatch({ type: SET_CURRENT_SORT, payload: SortOption.Balance });
-          const sortedWallets = state.tempOrder
-            .filter((item): item is WalletItem => item.type === ItemType.WalletSection)
-            .sort((a, b) => a.data.getBalance() - b.data.getBalance());
-          dispatch({ type: SET_TEMP_ORDER, payload: sortedWallets });
-          break;
-        }
-        case CommonToolTipActions.SortLabel.id: {
-          dispatch({ type: SET_CURRENT_SORT, payload: SortOption.Label });
-          const sortedWalletsByLabel = state.tempOrder
-            .filter((item): item is WalletItem => item.type === ItemType.WalletSection)
-            .sort((a, b) => a.data.getLabel()!.localeCompare(b.data.getLabel()!));
-          dispatch({ type: SET_TEMP_ORDER, payload: sortedWalletsByLabel });
-          break;
-        }
-        case CommonToolTipActions.MostRecentTransaction.id: {
-          dispatch({ type: SET_CURRENT_SORT, payload: SortOption.MostRecent });
-          const sortedWalletsByMostRecent = state.tempOrder
-            .filter((item): item is WalletItem => item.type === ItemType.WalletSection)
-            .sort((a, b) => {
-              return b.data.getTransactions()[0]?.time - a.data.getTransactions()[0]?.time;
-            });
-          dispatch({ type: SET_TEMP_ORDER, payload: sortedWalletsByMostRecent });
-          break;
-        }
-        case CommonToolTipActions.Reset.id: {
-          dispatch({ type: SET_TEMP_ORDER, payload: state.order });
-          break;
-        }
-      }
-    },
-    [state.order, state.tempOrder],
-  );
-
   const SaveButton = useMemo(
     () => <HeaderRightButton disabled={!hasUnsavedChanges} title={loc.send.input_done} onPress={handleClose} />,
     [handleClose, hasUnsavedChanges],
-  );
-
-  const MoreButton = useMemo(
-    () => <MoreOptionsButton isMenuPrimaryAction actions={moreOptionsActions} onPressMenuItem={moreOptionsOnPressMenuItem} />,
-    [moreOptionsActions, moreOptionsOnPressMenuItem],
-  );
-
-  const HeaderRight = useCallback(
-    () => (
-      <>
-        {MoreButton}
-        <View style={styles.separation} />
-        {SaveButton}
-      </>
-    ),
-    [MoreButton, SaveButton],
   );
 
   useLayoutEffect(() => {
@@ -417,10 +280,10 @@ const ManageWallets: React.FC = () => {
 
     setOptions({
       headerLeft: () => HeaderLeftButton,
-      headerRight: HeaderRight,
+      headerRight: () => SaveButton,
       headerSearchBarOptions: searchBarOptions,
     });
-  }, [setOptions, HeaderLeftButton, SaveButton, MoreButton, HeaderRight]);
+  }, [setOptions, HeaderLeftButton, SaveButton]);
 
   useFocusEffect(
     useCallback(() => {
@@ -713,8 +576,5 @@ const styles = StyleSheet.create({
   },
   dimmedText: {
     opacity: 0.8,
-  },
-  separation: {
-    width: 16,
   },
 });
