+++ typings/CommonToolTipActions.ts
@@ -50,9 +50,6 @@ const keys = {
   SortValue: 'sortValue',
   SortLabel: 'sortLabel',
   SortStatus: 'sortStatus',
-  SortBalance: 'sortBalance',
-  MostRecentTransaction: 'mostRecentTransaction',
-  Reset: 'reset',
 } as const;
 
 const icons = {
@@ -97,8 +94,6 @@ const icons = {
   ClearClipboard: { iconValue: 'clipboard' },
   SortASC: { iconValue: 'arrow.down.to.line' },
   SortDESC: { iconValue: 'arrow.up.to.line' },
-  MostRecentTransaction: { iconValue: 'clock' },
-  Reset: { iconValue: 'arrow.counterclockwise' },
 } as const;
 
 export const CommonToolTipActions = {
@@ -314,11 +309,6 @@ export const CommonToolTipActions = {
     id: keys.ResetToDefault,
     text: loc.settings.electrum_reset,
   },
-  Reset: {
-    id: keys.Reset,
-    text: loc.receive.reset,
-    Icon: icons.Reset,
-  },
   ClearHistory: {
     id: keys.ClearHistory,
     text: loc.settings.electrum_clear,
@@ -339,13 +329,11 @@ export const CommonToolTipActions = {
     id: keys.SortASC,
     text: loc.cc.sort_asc,
     icon: icons.SortASC,
-    keepsMenuPresented: true,
   },
   SortDESC: {
     id: keys.SortDESC,
     text: loc.cc.sort_desc,
     icon: icons.SortDESC,
-    keepsMenuPresented: true,
   },
   SortHeight: {
     id: keys.SortHeight,
@@ -362,15 +350,5 @@ export const CommonToolTipActions = {
   SortStatus: {
     id: keys.SortStatus,
     text: loc.cc.sort_status,
-  },
-  SortBalance: {
-    id: keys.SortBalance,
-    text: loc.wallets.balance,
-    icon: icons.ViewInBitcoin,
-  },
-  MostRecentTransaction: {
-    id: keys.MostRecentTransaction,
-    text: loc.transactions.details_title,
-    icon: icons.MostRecentTransaction,
   },
 } as const;
